# Modify cluster metadata